+++ src/globals.d.ts
@@ -1,3 +1,7 @@
+declare module 'real-react-dom/server' {
+  export function renderToStaticMarkup(element: any): string;
+}
+
 declare module '*.module.css' {
   const classnames: Record<string, string>;
   export default classnames;